Added interpereter

+++ stdlib/math.cpp
@@ -4,17 +4,17 @@
 
 namespace mipt_vm::stdlib {
 
-int Math::sqrt(int64_t val)
+double Math::sqrt(double val)
 {
     return std::sqrt(val);
 }
 
-int Math::sin(int64_t val)
+double Math::sin(double val)
 {
     return std::sin(val);
 }
 
-int Math::cos(int64_t val)
+double Math::cos(double val)
 {
     return std::cos(val);
 }
